refactor: 메인페이지 board 리팩토링
