+++ src/components/Main/MainFuction.tsx
@@ -6,35 +6,13 @@ import Navigation from '../../assets/Navigation.svg';
 import Calculator from '../../assets/Calculator.svg';
 import BookOpen from '../../assets/BookOpen.svg';
 import User from '../../assets/User.svg';
+import BoardsAtMain from './BoardAtMain';
 
 const progressState = [
   { id: 0, title: '원서 제출', date: '10/17~10/20' },
   { id: 1, title: '1차 발표', date: '10/24 18:00' },
   { id: 2, title: '원서 제출', date: '10/18 9:00' },
   { id: 3, title: '2차 발표', date: '11/03 10:00' },
-];
-
-const boardContent = [
-  {
-    id: 0,
-    title: '입학 공지사항',
-    content: [
-      '2023학년도 신입생 입학전 과제 제출 안내',
-      '2023학년도 신입생 오리엔테이션 집행 안내',
-      '2023학년도 신입생 대상 정밀 건강검진 안내',
-      '2023학년도 신입생 2차전형 안내',
-    ],
-  },
-  {
-    id: 1,
-    title: '자주 묻는 질문',
-    content: [
-      '전학 또는 편입으로 입학할 수 있나요?',
-      '자기소개서는 공백을 포함하나요?',
-      '다른 마이스터고에 중복 지원이 가능한가요?',
-      '여런 전형에 중복 지원이 가능한가요?',
-    ],
-  },
 ];
 
 interface ICurrentDate {
@@ -135,28 +113,7 @@ const MainFunction = () => {
             </_ShorcutButton>
           ))}
         </_Shortcut>
-        {boardContent.map((item) => (
-          <Board key={item.id}>
-            <BoardTitle>
-              <Text color="black600" size="title2">
-                {item.title}{' '}
-                <svg width="8" height="14" viewBox="0 0 8 14" fill="none" xmlns="http://www.w3.org/2000/svg">
-                  <path
-                    d="M7.53974 6.29004L1.87974 0.640037C1.78677 0.546308 1.67617 0.471914 1.55431 0.421145C1.43246 0.370377 1.30175 0.344238 1.16974 0.344238C1.03773 0.344238 0.90702 0.370377 0.785161 0.421145C0.663302 0.471914 0.552701 0.546308 0.459738 0.640037C0.273487 0.827399 0.168945 1.08085 0.168945 1.34504C0.168945 1.60922 0.273487 1.86267 0.459738 2.05004L5.40974 7.05004L0.459738 12C0.273487 12.1874 0.168945 12.4409 0.168945 12.705C0.168945 12.9692 0.273487 13.2227 0.459738 13.41C0.552352 13.5045 0.662797 13.5797 0.784673 13.6312C0.906548 13.6827 1.03743 13.7095 1.16974 13.71C1.30204 13.7095 1.43293 13.6827 1.5548 13.6312C1.67668 13.5797 1.78712 13.5045 1.87974 13.41L7.53974 7.76004C7.64124 7.66639 7.72225 7.55274 7.77766 7.42624C7.83307 7.29974 7.86167 7.16314 7.86167 7.02504C7.86167 6.88693 7.83307 6.75033 7.77766 6.62383C7.72225 6.49733 7.64124 6.38368 7.53974 6.29004Z"
-                    fill="#5F5F5F"
-                  />
-                </svg>
-              </Text>
-            </BoardTitle>
-            {item.content.map((item, i) => (
-              <BoardContent key={i}>
-                <Text color="realBlack" size="body1">
-                  {item.length > 20 ? `${item.slice(0, 20)}...` : item}
-                </Text>
-              </BoardContent>
-            ))}
-          </Board>
-        ))}
+        <BoardsAtMain />
       </_Discription>
     </_Wrapper>
   );
@@ -268,26 +225,4 @@ const _ShorcutButton = styled.div`
   height: 6rem;
   border: 1px solid #e6e6e6;
   border-radius: 5px;
-`;
-
-const Board = styled.div`
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 21rem;
-  height: 14rem;
-  margin-left: 2rem;
-`;
-
-const BoardTitle = styled.div`
-  width: 21rem;
-  padding-left: 4px;
-  padding-bottom: 8px;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.25);
-`;
-
-const BoardContent = styled.div`
-  width: 20.5rem;
-  height: 2rem;
-  border-bottom: 1px solid rgba(0, 0, 0, 0.1);
 `;
